Rename Rosetta parser to transactor
