+++ rosetta/transactor/convert.go
@@ -12,16 +12,13 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
-package transactions
+package transactor
 
 import (
+	sdk "github.com/onflow/flow-go-sdk"
 	"github.com/onflow/flow-go/model/flow"
-
-	"github.com/optakt/flow-dps/rosetta/identifier"
 )
 
-type Validator interface {
-	Account(rosAccountID identifier.Account) (address flow.Address, err error)
-	Block(rosBlockID identifier.Block) (height uint64, blockID flow.Identifier, err error)
-	Currency(currency identifier.Currency) (symbol string, decimals uint, err error)
+func convertAddress(address sdk.Address) flow.Address {
+	return flow.BytesToAddress(address[:])
 }
